Ai Quiz backend/requirement.txt
fastapi
uvicorn
python-docx
python-pptx
PyMuPDF
google-genai
python-multipart
python-dotenv